Added pytest and docstring
