+++ tests/store/api/conftest.py
@@ -4,6 +4,7 @@ import pytest
 from typing import Iterable, List
 
 from cg.constants import Pipeline
+from cg.meta.orders.pool_submitter import PoolSubmitter
 from cg.store import Store, models
 from cg.store.api.import_func import (
     parse_application_versions,
@@ -175,6 +176,55 @@ def fixture_rml_store(store: Store, helpers: StoreHelpers) -> Store:
         )
 
     return store
+
+
+@pytest.fixture(name="rml_pool_store")
+def fixture_rml_pool_store(case_id: str, helpers, store: Store, ticket_number: int):
+
+    customer_group = store.add_customer_group("dummy_group", "dummy group")
+    new_customer = store.add_customer(
+        internal_id="cust000",
+        name="Test customer",
+        scout_access=True,
+        customer_group=customer_group,
+        invoice_address="skolgatan 15",
+        invoice_reference="abc",
+    )
+    store.add_commit(new_customer)
+
+    application = store.add_application(
+        tag="RMLP05R800",
+        category="rml",
+        description="Ready-made",
+        percent_kth=80,
+        percent_reads_guaranteed=75,
+        sequencing_depth=0,
+    )
+    store.add_commit(application)
+
+    app_version = store.add_version(
+        application=application,
+        version=1,
+        valid_from=dt.datetime.today(),
+        prices={"standard": 12, "priority": 222, "express": 123, "research": 12},
+    )
+    store.add_commit(app_version)
+
+    new_pool = store.add_pool(
+        customer=new_customer,
+        name="Test",
+        order="Test",
+        ordered=dt.datetime.today(),
+        application_version=app_version,
+    )
+    store.add_commit(new_pool)
+    new_case = helpers.add_case(
+        store=store,
+        internal_id=case_id,
+        name=PoolSubmitter.create_case_name(ticket=ticket_number, pool_name="Test"),
+    )
+    store.add_commit(new_case)
+    yield store
 
 
 @pytest.fixture(name="re_sequenced_sample_store")
